requirements.txt
confluent-kafka[avro]==2.2.0
pyspark==3.4.1
apache-airflow==2.7.1
delta-spark==2.4.0
great-expectations==0.17.15
prometheus-client==0.17.1
pytest==7.4.0
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0
locust==2.15.1
dbt-core==1.6.1
dbt-postgres==1.6.1
feast==0.33.0
redis==5.0.0
openlineage-python==1.5.0
